+++ src/main/java/com/example/imc/Controllers/OrderController.java
@@ -7,14 +7,12 @@ import javafx.animation.FadeTransition;
 import javafx.application.Platform;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.fxml.FXML;
-import javafx.scene.control.Alert;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.effect.BoxBlur;
 import javafx.scene.input.KeyCode;
 import javafx.scene.layout.Pane;
 import javafx.scene.text.Text;
+import javafx.stage.StageStyle;
 import javafx.util.Duration;
 
 import java.sql.ResultSet;
@@ -23,6 +21,7 @@ import java.sql.Statement;
 import java.time.LocalDateTime;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Optional;
 
 public class OrderController {
     QueryHandler queryHandler = new QueryHandler();
@@ -47,6 +46,20 @@ public class OrderController {
     TextField quantityController;
 
     @FXML
+    Text totalOrders;
+    @FXML
+    Text totalSold;
+    @FXML
+    Text revenue;
+    @FXML
+    Text productName;
+    @FXML
+    Text totalQuantity;
+    @FXML
+    Text lastOrderedDate;
+    @FXML
+    Text lastOrderedProduct;
+    @FXML
     private Pane mainPane;
 
     @FXML
@@ -58,14 +71,57 @@ public class OrderController {
     public void initialize() throws SQLException {
         stmt = DatabaseHandler.getStatement();
 
-        ResultSet rs = stmt.executeQuery("select * from orders");
+        // Total Orders
+        ResultSet resultSet = stmt.executeQuery("SELECT COUNT(*) FROM Orders;");
+        resultSet.next();
+        int totalOrders = resultSet.getInt(1);
+        this.totalOrders.setText(String.valueOf(totalOrders));
+        // Products Sold
+        resultSet = stmt.executeQuery("SELECT SUM(OrderQuantity) FROM Orders;");
+        resultSet.next();
+        int totalSold = resultSet.getInt(1);
+        this.totalSold.setText(String.valueOf(totalSold));
+        // Revenue
+        resultSet = stmt.executeQuery("SELECT SUM(p.ProductPrice * o.OrderQuantity) " +
+                "FROM Orders o INNER JOIN Products p ON o.ProductID = p.ProductID;");
+        resultSet.next();
+        double revenue = resultSet.getDouble(1);
+        this.revenue.setText(String.valueOf(revenue));
 
-        while (rs.next()) {
-            String orderID = rs.getString("OrderID");
-            String productID = rs.getString("ProductID");
-            String quantity = rs.getString("OrderQuantity");
-            LocalDateTime dateTime = rs.getTimestamp("OrderDate").toLocalDateTime();
-            Platform.runLater(() -> addOrder(orderID, productID, quantity, dateTime.toString()));
+        resultSet = stmt.executeQuery("SELECT p.ProductName, o.OrderDate " +
+                "FROM Orders o JOIN Products p ON o.ProductID = p.ProductID " +
+                "WHERE o.OrderDate = (SELECT MAX(OrderDate) FROM Orders);");
+
+        if (resultSet.next()) {
+            String lastOrderedProduct = resultSet.getString("ProductName");
+            lastOrderedProduct = lastOrderedProduct.length() > 6 ? lastOrderedProduct.substring(0, 6) + "..." : lastOrderedProduct;
+
+            Date lastOrderDate = resultSet.getDate("OrderDate");
+            this.lastOrderedProduct.setText(lastOrderedProduct);
+            this.lastOrderedDate.setText(lastOrderDate.toString());
+        }
+
+        resultSet = stmt.executeQuery("SELECT p.ProductName, SUM(o.OrderQuantity) AS TotalQuantity " +
+                "FROM Orders o JOIN Products p ON o.ProductID = p.ProductID " +
+                "GROUP BY p.ProductName " +
+                "ORDER BY TotalQuantity DESC " +
+                "LIMIT 1;");
+
+        if (resultSet.next()) {
+            String productName = resultSet.getString("ProductName");
+            productName = productName.length() > 6 ? productName.substring(0, 6) + "..." : productName;
+            int totalQuantity = resultSet.getInt("TotalQuantity");
+            this.productName.setText(productName);
+            this.totalQuantity.setText(String.valueOf(totalQuantity));
+        }
+        resultSet = stmt.executeQuery("select * from orders");
+
+        while (resultSet.next()) {
+            String orderID = resultSet.getString("OrderID");
+            String productID = resultSet.getString("ProductID");
+            String quantity = resultSet.getString("OrderQuantity");
+            String orderDate = resultSet.getString("OrderDate");
+            Platform.runLater(() -> addOrder(orderID, productID, quantity, orderDate));
         }
         // Add event listener for delete key press
         tableView.setOnKeyPressed(event -> {
@@ -76,23 +132,30 @@ public class OrderController {
                     // Call a method to delete the row from the database
                     String orderID = selectedOrder.getOrderID();
 
-                    boolean status = deleteFromDatabase(orderID);
+                    // Create a confirmation dialog
+                    Alert confirmationDialog = new Alert(Alert.AlertType.CONFIRMATION);
+                    confirmationDialog.setTitle("Confirmation");
+                    confirmationDialog.setHeaderText(null);
+                    confirmationDialog.setContentText("Are you sure you want to delete this order from the database?");
+                    confirmationDialog.initStyle(StageStyle.UNDECORATED); // Optional: Removes the default window decorations
 
-                    if (status)
-                    {
-                        tableView.getItems().remove(selectedOrder);
+                    // Show the confirmation dialog and wait for user response
+                    Optional<ButtonType> result = confirmationDialog.showAndWait();
 
+                    // Check if the user confirmed the deletion
+                    if (result.isPresent() && result.get() == ButtonType.OK) {
+                        boolean status = deleteFromDatabase(orderID);
+                        if (status) {
+                            tableView.getItems().remove(selectedOrder);
+                        } else {
+                            String errorMessage = "Error deleting the order from the database";
+                            Alert alert = new Alert(Alert.AlertType.ERROR);
+                            alert.setTitle("Error");
+                            alert.setHeaderText(null);
+                            alert.setContentText(errorMessage);
+                            alert.showAndWait();
+                        }
                     }
-                    else
-                    {
-                        String errorMessage = "Error deleting the order from the database";
-                        Alert alert = new Alert(Alert.AlertType.ERROR);
-                        alert.setTitle("Error");
-                        alert.setHeaderText(null);
-                        alert.setContentText(errorMessage);
-                        alert.showAndWait();
-                    }
-
                     // Remove the product from the TableView
                 }
             }
@@ -100,7 +163,7 @@ public class OrderController {
     }
 
     @FXML
-    void kAddOrder() {
+    private void kAddOrder() {
         // Apply a BoxBlur effect to the mainPane to make it blur
         BoxBlur blur = new BoxBlur(5, 5, 3);
         mainPane.setEffect(blur);
@@ -125,13 +188,10 @@ public class OrderController {
         Date currentDate = Calendar.getInstance().getTime();
         boolean status = queryHandler.insertOrder(order.getOrderID(), order.getOrderDate(currentDate), order.getProductID(), order.getOrderQuantity());
         // Add the order to the UI
-        if (status)
-        {
+        if (status) {
             onErrorText.setVisible(false);
             addOrder(order.getOrderID(), order.getProductID(), order.getOrderQuantity(), order.getOrderDate(currentDate));
-        }
-
-        else {
+        } else {
             onErrorText.setVisible(true);
             return;
         }
@@ -140,7 +200,6 @@ public class OrderController {
         fadeOutTransition.setToValue(0);
         fadeOutTransition.setOnFinished(event -> popupPane.setVisible(false));
         fadeOutTransition.play();
-
         // Remove the BoxBlur effect from the mainPane
         mainPane.setEffect(null);
     }
@@ -154,7 +213,6 @@ public class OrderController {
         fadeOutTransition.setToValue(0);
         fadeOutTransition.setOnFinished(event -> popupPane.setVisible(false));
         fadeOutTransition.play();
-
         // Remove the BoxBlur effect from the mainPane
         mainPane.setEffect(null);
     }
@@ -165,12 +223,7 @@ public class OrderController {
         c1.setCellValueFactory(cellData -> cellData.getValue().orderIDProperty());
         c2.setCellValueFactory(cellData -> cellData.getValue().productIDProperty());
         c3.setCellValueFactory(cellData -> cellData.getValue().orderQuantityProperty());
-        c4.setCellValueFactory(cellData -> {
-            Date currentDate = Calendar.getInstance().getTime();
-
-            String date = cellData.getValue().getOrderDate(currentDate);
-            return new SimpleStringProperty(date);
-        });
+        c4.setCellValueFactory(cellData -> new SimpleStringProperty(dateTime.split(" ")[0]));
         // Add the custom row to the table
         tableView.getItems().add(order);
 
@@ -184,7 +237,6 @@ public class OrderController {
         } catch (SQLException e) {
             e.printStackTrace();
             return false;
-            // Handle any exception that occurs during the database operation
         }
     }
 }
